First version of MCMC chain for Domino
This program uses an MCMC chain
to  sample from the distribution of
a generalized domino tiling.

The generalized domino have 3 types of
blocks: 4 by 1, 1 by 4, and 2 by 2.

We are interested in a “random tiling”, sampled from the distribution
on tilings
that gives equal probability to each tiling.

To ensure that MCMC chain converges, with start from two opposite
tilings (all tiles horizontal and all tiles vertical) and run the MCMC
on both tilings until they couple.

The deficiency of this variant of the chain is that it appears that its
running time is
exponential in the number of tiles (i.e., the size of the board.)